Start to make necessary changes for card system to `Entity` class.
FEAT: a new optional parameter `crd` is added to `Entity` class's constructor. It is a tuple returned by one of the methods in `Card` and contains 4 stats of the Entity (ms, luck, etc.) If it exists (not None), it comes into effect and overrides the 4 class members of `Entity` no matter they were already assigned to or not.

Additional changes: minor changes to documentations of `Entity` `Card`